Moved opal_filter.rb to spec/rspec/shared. Completely removed lib-opal/opal/rspec/upstream-specs-support directory.
