+++ spec/rspec/shared/opal_filters.rb
